+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.14.0] - 2023-07-27
+### Added
+- [@mjkuranda](https://github.com/mjkuranda): GameAnimation as a game resource.
+- [@mjkuranda](https://github.com/mjkuranda): Spawning an explosion animation after destroying asteroid.
+- [@mjkuranda](https://github.com/mjkuranda): Damaging player spaceship.
+
+### Changed
+- [@mjkuranda](https://github.com/mjkuranda): RetroRenderer refactored rendering entities.
+- [@mjkuranda](https://github.com/mjkuranda): Missiles damage other entities instead of immediately destroy them.
+
 ## [0.13.0] - 2023-07-26
 ### Added
 - [@mjkuranda](https://github.com/mjkuranda): Each entity has an image reference.